# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Essa é a API de Controle de Caixa, que pode ser utilizada junto com a aplicação para controle de caixa de pequenos négocios",
        "version": "1.0.1",
        "title": "Controle de Caixa API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "email": "carleonepsouza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "https://app-loja.azurewebsites.net/v1/api"
        }
    ],
    "host": "https://app-loja.azurewebsites.net/v1/api",
...
    "basePath": "/v1",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/orders": {
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Register Order",
                "description": "This route will be responsible for register an order",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            },
                            "examples": {
                                "sale": {
                                    "value": {
                                        "produtos": {
                                            "type":"array",
                                            "$ref":"#/components/schemas/Product"
                                        },
                                        "nvenda": 3,
                                        "total": 24.0,
                                        "formaPagamnto": "Credito",
                                        "troco": 0,
                                        "valorPago": 24.0,
                                        "user": "640a0d0a9b99632ae8b84ab0",
                                        "status": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "409": {
                        "description": "Conflit!"
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Return a list of orders",
                "description": "This route will be responsible for return a list of orders",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Return a specific order",
                "description": "This route will be responsible for return an order by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "orders"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated order object",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Order"
                        }
                    }
                ],
                "summary": "Update a specific order",
                "description": "This route will be responsible for update an order by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            },
                            "examples": {
                                "sale": {
                                    "value": {
                                        "product": "Cerveja Devassa",
                                        "quantity": 3,
                                        "total": 24.0,
                                        "paymentMethod": "Credito",
                                        "change": 0,
                                        "paymentValue": 24.0,
                                        "user": "",
                                        "status": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "orders"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "summary": "Delete a specific order",
                "description": "This route will be responsible for delete an order by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Sorry, Internal Server Error!"
                    },
                    "400": {
                        "description": "Bad Request!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "value": {
                        "type": "number"
                    },
                    "classification": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "volume": {
                        "type": "number"
                    }
                }
            },
            "Category": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "number"
                    },
                    "cpfCnpj": {
                        "type": "number"
                    },
                    "password": {
                        "type": "string"
                    },
                    "apiKey": {
                        "type": "string"
                    },
                    "address": {
                        "$ref": "#/components/schemas/Address"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
            "Address": {
                "type": "object",
                "properties": {
                    "street": {
                        "type": "string"
                    },
                    "neighborhood": {
                        "type": "string"
                    },
                    "zipCode": {
                        "type": "number"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
            "Profile": {
                "type": "object",
                "properties": {
                    "role": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
            "Order": {
                "type": "object",
                "properties": {
                    "nvenda": {
                        "type": "number"
                    },
                    "total": {
                        "type": "number"
                    },
                    "formaPagamnto": {
                        "type": "string"
                    },
                    "troco": {
                        "type": "number"
                    },
                    "valorPago": {
                        "type": "number"
                    },
                    "produtos": {
                        "type":"array",
                        "$ref": "#/components/schemas/Product"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "status": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}